LPS-56789 Rename module
