+++ modules/apps/portlet-display-template/portlet-display-template/.project
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <projectDescription>
-	<name>portlet-display-template-service-master</name>
+	<name>portlet-display-template-master</name>
 	<comment></comment>
 	<projects></projects>
 	<buildSpec>
